Generalised function 'make_handlers'
Function 'make_handlers' can now take an arbitrary number of
key codes to handle. The key states of each key are kept track of
in the associated 'key_states' array (which is available to the
updating function 'f'). This is indexed by key code.

e.g. 'key_states[83]' returns the key state of key code '83'.

Since the 'value' that is manipulated by the key handlers is now of
arbitrary form this allows us to create arbitrarily complex key
handlers.

For instance we now control the rotation of the object using four keys,
'W', 'A', 'S', and 'D'. 'W' and 'S' control the "phi" displacement/velocity
and 'A' and 'D' control the "theta" displacement/velocity.