Merged key-handler.julius and mouse-handler.js
We now just have one module for input handling. I did this because
pan can be controlled by either the keys or the mouse, but it is
manipulating the same internal state.

Future work: Add a function to return the internal state so that when
we resize the canvas we can pull out this state and feed it back into
fresh handlers.